+++ src/GeekBrains/Level1/Lesson8/Homework/TicTacToe.java
@@ -1,0 +1,7 @@
+package GeekBrains.Level1.Lesson8.Homework;
+
+public class TicTacToe {
+    public static void main(String[] args) {
+        new GameWindow();
+    }
+}
