+++ src/GeekBrains/Level2/Lesson1/Homework/Task1.java
@@ -1,0 +1,7 @@
+package GeekBrains.Level2.Lesson1.Homework;
+
+public class Task1 {
+    public static void main(String[] args) {
+
+    }
+}
